Moves trauma damage code into a different file
